+++ src/capabilities/get-monitored-entity-details.ts
@@ -31,19 +31,19 @@ export const getMonitoredEntityDetails = async (
   const dqlResponse = await executeDql(dtClient, { query: dql });
 
   // verify response and length
-  if (!dqlResponse || dqlResponse.length === 0) {
+  if (!dqlResponse || !dqlResponse.records || dqlResponse.records.length === 0) {
     console.error(`No entity found for ID: ${entityId}`);
     return;
   }
 
   // in case we have more than one entity -> log it
-  if (dqlResponse.length > 1) {
+  if (dqlResponse.records.length > 1) {
     console.error(
-      `Multiple entities (${dqlResponse.length}) found for entity ID: ${entityId}. Returning the first one.`,
+      `Multiple entities (${dqlResponse.records.length}) found for entity ID: ${entityId}. Returning the first one.`,
     );
   }
 
-  const entity = dqlResponse[0];
+  const entity = dqlResponse.records[0];
   // make typescript happy; entity should never be null though
   if (!entity) {
     console.error(`No entity found for ID: ${entityId}`);